requirements.txt
streamlit
phidata
qdrant-client
openai
pypdf
python-docx
rich
typer
